Adicionando relatório.
